Refactor mouse impl to HidTable
